refactor: make GetPrincipal reusable
move it to location usable from request handlers.
