+++ internal/infrastructure/middleware/enforce_identity_test.go
@@ -16,7 +16,6 @@ import (
 	"github.com/podengo-project/idmsvc-backend/internal/api/header"
 	internal_errors "github.com/podengo-project/idmsvc-backend/internal/errors"
 	"github.com/podengo-project/idmsvc-backend/internal/test/builder/api"
-	builder_api "github.com/podengo-project/idmsvc-backend/internal/test/builder/api"
 	identity "github.com/redhatinsights/platform-go-middlewares/v2/identity"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -677,68 +676,5 @@ func TestEnforceServiceAccountPredicate(t *testing.T) {
 		} else {
 			assert.NoError(t, err)
 		}
-	}
-}
-
-func TestGetPrincipal(t *testing.T) {
-	type TestCase struct {
-		Name     string
-		Given    identity.XRHID
-		Expected string
-	}
-	testCases := []TestCase{
-		{
-			Name:     "User identity",
-			Given:    builder_api.NewUserXRHID().WithUserID("user-id").Build(),
-			Expected: "user-id",
-		},
-		{
-			Name:     "System identity",
-			Given:    builder_api.NewSystemXRHID().WithCommonName("system-id").Build(),
-			Expected: "system-id",
-		},
-		{
-			Name:     "Service Account identity",
-			Given:    builder_api.NewServiceAccountXRHID().WithClientID("service-account-id").Build(),
-			Expected: "service-account-id",
-		},
-		{
-			Name: "X509 identity",
-			Given: identity.XRHID{
-				Identity: identity.Identity{
-					Type: identityTypeX509,
-					X509: &identity.X509{
-						SubjectDN: "CN=cert identity, O=domain.test",
-					},
-				},
-			},
-			Expected: "CN=cert identity, O=domain.test",
-		},
-		{
-			Name: "Red Hat UUID",
-			Given: identity.XRHID{
-				Identity: identity.Identity{
-					Type: identityTypeAssociate,
-					Associate: &identity.Associate{
-						RHatUUID: "redhat-id",
-					},
-				},
-			},
-			Expected: "redhat-id",
-		},
-		{
-			Name: "Unknown identity",
-			Given: identity.XRHID{
-				Identity: identity.Identity{
-					Type: identityTypeUnknown,
-				},
-			},
-			Expected: identityTypeUnknown,
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Log(testCase.Name)
-		require.Equal(t, testCase.Expected, getPrincipal(&testCase.Given))
 	}
 }
